v4.0 A.2: split openai_compat.py — leaf-first per GRAEAE
Extract translation layer from the 1366-line route file into a
proper domain package. Route file becomes thin (HTTP shim);
domain holds the schemas, providers, content helpers, streaming,
and orchestration; db holds the SQL.

- mnemos/api/routes/openai_compat.py: thinned to route shim
  (parse request, call domain, return response).
- mnemos/domain/openai_compat/{schemas, content, providers,
  streaming, router}.py: business logic per concern.
- mnemos/db/openai_compat_repo.py: SQL extracted from
  _search_mnemos_context and _get_model_recommendation
  (where applicable).
- Imports updated across the codebase.

No wire-shape changes. No behavior changes. 925 unit tests pass.

Authored-By: Codex
