+++ mnemos/api/routes/sessions.py
@@ -13,8 +13,9 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 
 import mnemos.core.lifecycle as _lc
 from mnemos.api.dependencies import UserContext, get_current_user
-from mnemos.api.routes.openai_compat import _route_to_provider, _search_mnemos_context
 from mnemos.core.security import scope_namespace
+from mnemos.db.openai_compat_repo import fetch_memory_context as _search_mnemos_context
+from mnemos.domain.openai_compat.providers import _route_to_provider
 from mnemos.domain.models import (
     ChatMessage,
     SessionContext,
