+++ mnemos/domain/openai_compat/__init__.py
@@ -1,0 +1,44 @@
+from .providers import MODEL_ALIASES, TASK_CAPABILITY_MAP, OpenAICompatError
+from .router import StreamingChatCompletion, chat_completion, get_model, list_models
+from .schemas import (
+    ChatCompletionChoice,
+    ChatCompletionDelta,
+    ChatCompletionRequest,
+    ChatCompletionResponse,
+    ChatCompletionResponseMessage,
+    ChatCompletionStreamChoice,
+    ChatCompletionStreamRequest,
+    ChatCompletionStreamResponse,
+    ChatContent,
+    ChatMessage,
+    ContentBlock,
+    ModelInfo,
+    ModelsResponse,
+    Tool,
+    ToolFunction,
+)
+
+__all__ = [
+    "MODEL_ALIASES",
+    "TASK_CAPABILITY_MAP",
+    "OpenAICompatError",
+    "StreamingChatCompletion",
+    "chat_completion",
+    "get_model",
+    "list_models",
+    "ChatCompletionChoice",
+    "ChatCompletionDelta",
+    "ChatCompletionRequest",
+    "ChatCompletionResponse",
+    "ChatCompletionResponseMessage",
+    "ChatCompletionStreamChoice",
+    "ChatCompletionStreamRequest",
+    "ChatCompletionStreamResponse",
+    "ChatContent",
+    "ChatMessage",
+    "ContentBlock",
+    "ModelInfo",
+    "ModelsResponse",
+    "Tool",
+    "ToolFunction",
+]
